+++ web/src/stores/app-store.ts
@@ -395,34 +395,7 @@ export const useAppStore = defineStore('app', () => {
   }
 
   const migrateToPlannerTabs = (data: BackendFactoryDataResponse): void => {
-    if (!data.tabs) {
-      // Migration is required
-      console.log('migrateToTabs: Migration required.')
-      // @ts-ignore
-      let factories = data.data as Factory[] | undefined
 
-      if (!factories) {
-        console.error('migrateToTabs: No factories found to migrate.')
-        factories = []
-      }
-
-      // Migrate factories to tabs
-
-      const defaultTab: FactoryTab = {
-        id: 'default',
-        name: 'Default',
-        factories,
-      }
-
-      const newData = {
-        user: data.user,
-        tabs: [defaultTab],
-        lastSaved: data.lastSaved,
-      }
-
-      // Set the new data to the planner
-      this.appStore.setTabs(newData)
-    }
   }
 
   return {
