+++ web/src/stores/sync-store.spec.ts
@@ -189,4 +189,11 @@ describe('useSyncStore', () => {
       expect(syncStore.stopSyncing.value).toBe(true)
     })
   })
+
+  describe('migrateToTabSync', () => {
+    it('should handle migration to tab sync', () => {
+      const data = { data: 'mock-data' }
+      syncStore.migrateToTabSync(data)
+    }
+  })
 })
